worker/requirements.txt
psycopg2-binary==2.9.9
minio==7.2.7
requests==2.32.3
python-dotenv==1.0.1
